Fix conflict with upstream change for Glance RBD
Some time back during crunch time before a release, we discovered
that we needed to add params to the openstack::controller class
in order to make Ceph-backed Glance work correctly. Due to the lack
of time to get the patch pushed upstream, the patch was applied
locally with the intent of sending it upstream later. However,
that somehow slipped through the cracks and when upstream later
added the same parameters, they did so using a slightly different
name. In order to eliminate the local patch, we need to change
core.pp to use upstream's version of the patch. That was worked out in
the last commit, but one part of the upstream change didn't cause
a merge conflict and slipped through.  It leads to a catalog failure due
to duplicate parameters, so this commit fixes it.

Closes-Bug: #1229176